Define NULL as __nullptr for C++.
This is non-standard by a slim technicality - stringifying NULL in a
macro can show it is not a zero value. However, Clang's sentinel
detection requires than the sentinel value be a pointer type. This means
Musl's use of 0L for C++ NULL does not work with -Wsentinel. Since
everyone else uses __null, we may as well be wrong with good company.

Bug: 110161494
Change-Id: I9079041f276b0da0f6324b6b67966852f140f631